v1.13.1 - rework many things that now work differently - among those all things involving items, especially class definitions

+++ src/net/slipcor/pvparena/arena/PlayerState.java
@@ -242,7 +242,7 @@ public final class PlayerState {
      * @param player the player to set
      * @param value  the health value
      */
-    public static void playersetHealth(final Player player, final int value) {
+    public static void playersetHealth(final Player player, final double value) {
         debug.i("setting health to " + value + "/20", player);
         if (Bukkit.getServer().getPluginManager().getPlugin("Heroes") == null) {
             player.setHealth(value);
